Aggiunto Dapper Plus e importazione listino da txt

+++ Database/Models/Mamg0.cs
@@ -4,8 +4,8 @@ namespace Database.Models
     public class Mamg0
     {
         public float Pezzo { get; set; } = 0;
-        public float PrezzoListino { get; set; } = 0;
-        public float PrezzoNetto { get; set; } = 0;
+        public decimal PrezzoListino { get; set; } = 0;
+        public decimal PrezzoNetto { get; set; } = 0;
         public float Sconto1 { get; set; } = 0;
         public float Sconto2 { get; set; } = 0;
         public float Sconto3 { get; set; } = 0;
